--- a/src/lib/logical_query_plan/lqp_translator.cpp
+++ b/src/lib/logical_query_plan/lqp_translator.cpp
@@ -517,8 +517,8 @@ std::vector<std::shared_ptr<AbstractExpression>> LQPTranslator::_translate_expre
         return ExpressionVisitation::DoNotVisitArguments;
       }
 
-      Assert(expression->type != ExpressionType::LQPColumn,
-             "Failed to resolve Column '"s + expression->as_column_name() + "', LQP is invalid");
+      AssertInput(expression->type != ExpressionType::LQPColumn,
+                  "Failed to resolve Column '"s + expression->as_column_name() + "', LQP is invalid");
 
       return ExpressionVisitation::VisitArguments;
     });