--- a/src/lib/storage/frame_of_reference/frame_of_reference_iterable.hpp
+++ b/src/lib/storage/frame_of_reference/frame_of_reference_iterable.hpp
@@ -43,6 +43,8 @@ class FrameOfReferenceIterable : public PointAccessibleColumnIterable<FrameOfRef
     });
   }
 
+  size_t _on_size() const { return _column.size(); }
+
  private:
   const FrameOfReferenceColumn<T>& _column;