--- a/src/lib/storage/reference_column/reference_column_iterable.hpp
+++ b/src/lib/storage/reference_column/reference_column_iterable.hpp
@@ -28,6 +28,8 @@ class ReferenceColumnIterable : public ColumnIterable<ReferenceColumnIterable<T>
     functor(begin, end);
   }
 
+  size_t _on_size() const { return _column.size(); }
+
  private:
   const ReferenceColumn& _column;