--- a/src/lib/storage/run_length_column/run_length_column_iterable.hpp
+++ b/src/lib/storage/run_length_column/run_length_column_iterable.hpp
@@ -33,6 +33,8 @@ class RunLengthColumnIterable : public PointAccessibleColumnIterable<RunLengthCo
     functor(begin, end);
   }
 
+  size_t _on_size() const { return _column.size(); }
+
  private:
   const RunLengthColumn<T>& _column;