--- a/src/lib/storage/value_column/value_column_iterable.hpp
+++ b/src/lib/storage/value_column/value_column_iterable.hpp
@@ -40,6 +40,8 @@ class ValueColumnIterable : public PointAccessibleColumnIterable<ValueColumnIter
     }
   }
 
+  size_t _on_size() const { return _column.size(); }
+
  private:
   const ValueColumn<T>& _column;