--- a/src/test/tables/aggregateoperator/groupby_int_3gb_0agg/count_star.tbl
+++ b/src/test/tables/aggregateoperator/groupby_int_3gb_0agg/count_star.tbl
@@ -1,0 +1,13 @@
+a|c|d|COUNT(*)
+int_null|int_null|int_null|long
+12345|20|null|1
+12345|24|5|1
+12345|30|null|1
+12345|33|5|3
+12345|null|2|2
+123|20|3|1
+123|20|5|1
+123|null|null|1
+12|10|null|2
+12|null|2|1
+null|null|null|2