--- a/src/test/tables/aggregateoperator/groupby_int_3gb_1agg/sum_filtered.tbl
+++ b/src/test/tables/aggregateoperator/groupby_int_3gb_1agg/sum_filtered.tbl
@@ -1,0 +1,4 @@
+a|b|d|SUM(c)
+int|float|int|long
+12|350.7|1|10
+12|350.7|4|10