… src/test/tables/aggregateoperator/groupby_int_3gb_0agg/input.tbl
… src/test/tables/aggregateoperator/groupby_int_3gb_0agg/input_null.tbl
… src/test/tables/aggregateoperator/groupby_int_3gb_0agg/result.tbl
… src/test/tables/aggregateoperator/groupby_int_3gb_1agg/avg.tbl
… src/test/tables/aggregateoperator/groupby_int_3gb_1agg/count.tbl
… src/test/tables/aggregateoperator/groupby_int_3gb_1agg/input.tbl
… src/test/tables/aggregateoperator/groupby_int_3gb_1agg/max.tbl
… src/test/tables/aggregateoperator/groupby_int_3gb_1agg/min.tbl
… src/test/tables/aggregateoperator/groupby_int_3gb_1agg/sum.tbl