Couple of performance improvements. fixes #1025 

* Preallocate vectors for materialization 
* Specializations for GROUP BY with one and two columns 
*  Aggregate: Retrieve results[hash_keys[chunk_offset]] only once 

```
+-----------+------------------+------+------------------+------+--------+---------------------------------+
| Benchmark | prev. iter/s     | runs | new iter/s       | runs | change | p-value (significant if <0.001) |
+-----------+------------------+------+------------------+------+--------+---------------------------------+
| TPC-H 1   | 0.803068459034   | 49   | 1.65015637875    | 100  | +105%  |                          0.0000 |
| TPC-H 2   | 17.2374401093    | 1000 | 17.3384361267    | 1000 | +1%    |     (run time too short) 0.0000 |
| TPC-H 3   | 11.2716722488    | 677  | 11.4899930954    | 690  | +2%    |                          0.0000 |
| TPC-H 4   | 0.130990862846   | 8    | 0.131025746465   | 8    | +0%    |        (not enough runs) 0.5956 |
| TPC-H 5   | 7.1162571907     | 427  | 7.111140728      | 427  | -0%    |                          0.0049 |
| TPC-H 6   | 35.9103965759    | 1000 | 38.2005577087    | 1000 | +6%    |     (run time too short) 0.0000 |
| TPC-H 7   | 2.63078999519    | 158  | 2.71895647049    | 164  | +3%    |                          0.0000 |
| TPC-H 8   | 2.11616611481    | 127  | 2.09168195724    | 126  | -1%    |                          0.0095 |
| TPC-H 9   | 1.20813333988    | 73   | 1.22442007065    | 74   | +1%    |                          0.0000 |
| TPC-H 10  | 8.04513454437    | 483  | 8.57305812836    | 515  | +7%    |                          0.0000 |
| TPC-H 11  | 20.4952487946    | 1000 | 21.3759651184    | 1000 | +4%    |     (run time too short) 0.0000 |
| TPC-H 12  | 12.4446716309    | 747  | 12.9786014557    | 779  | +4%    |                          0.0000 |
| TPC-H 13  | 7.2131023407     | 433  | 9.39429187775    | 564  | +30%   |                          0.0000 |
| TPC-H 14  | 27.9485912323    | 1000 | 29.362651825     | 1000 | +5%    |     (run time too short) 0.0000 |
| TPC-H 15  | 0.142395094037   | 9    | 0.20015373826    | 13   | +41%   |        (not enough runs) 0.0000 |
| TPC-H 16  | 4.71243143082    | 283  | 4.99194812775    | 300  | +6%    |                          0.0000 |
| TPC-H 17  | 1.1221511364     | 68   | 1.11851346493    | 68   | -0%    |                          0.0000 |
| TPC-H 18  | 0.00699127931148 | 1    | 0.0226565748453  | 2    | +224%  |           (not enough runs) nan |
| TPC-H 19  | 1.11460006237    | 67   | 1.17799425125    | 71   | +6%    |                          0.0000 |
| TPC-H 20  | 0.00466411514208 | 1    | 0.00466101476923 | 1    | -0%    |           (not enough runs) nan |
| TPC-H 21  | 0.0788900107145  | 5    | 0.0788913741708  | 5    | +0%    |        (not enough runs) 0.8897 |
| TPC-H 22  | 0.254700601101   | 16   | 0.255093455315   | 16   | +0%    |                          0.0000 |
| average   |                  |      |                  |      | +20%   |                                 |
+-----------+------------------+------+------------------+------+--------+---------------------------------+
```

Full results (gcc/clang) here: https://gist.github.com/mrks/21a687801eea947675bdae9673d48df9